moved extract-choices-fragment-json to "helpers_server" submodule in tests to prevent ns collisions
